separating population control from walkers. walkers are nicely cleaned up.
