+++ ipie/walkers/uhf_walkers.py
@@ -31,16 +31,16 @@ from ipie.trial_wavefunction.wavefunction_base import TrialWavefunctionBase
 def get_initial_walker(trial: TrialWavefunctionBase)->(int,numpy.ndarray):
     if isinstance(trial, SingleDet):
         initial_walker = trial.psi
-        ndets = 1
+        num_dets = 1
     elif isinstance(trial, ParticleHoleWicks):
         initial_walker = numpy.hstack([trial.psi0a, trial.psi0b])
-        ndets = trial.num_dets
+        num_dets = trial.num_dets
     elif isinstance(trial, ParticleHoleWicksNonChunked):
         initial_walker = numpy.hstack([trial.psi0a, trial.psi0b])
-        ndets = trial.num_dets
+        num_dets = trial.num_dets
     else:
         raise Exception("Unrecognized trial type in get_initial_walker")
-    return ndets, initial_walker
+    return num_dets, initial_walker
 
 class UHFWalkers(BaseWalkers):
     """UHF style walker.
@@ -55,30 +55,18 @@ class UHFWalkers(BaseWalkers):
         self,
         initial_walker:numpy.ndarray,
         nup:int, ndown:int, nbasis:int,
-        num_walkers_local:int,
-        num_walkers_global:int,
-        num_steps:int,
-        ndets:int = None, # it's here for interface purposes
+        nwalkers:int,
         mpi_handler=None,
-        pop_control_method="pair_branch",
-        min_weight:float=0.1,
-        max_weight:float=4,
-        reconfiguration_frequency:int=50,
         verbose:bool=False
     ):
         assert len(initial_walker.shape) == 2
+        self.nup = nup
+        self.ndown = ndown
         self.nbasis = nbasis
+        self.mpi_handler = mpi_handler
 
         super().__init__(
-            nup, ndown,
-            num_walkers_local,
-            num_walkers_global,
-            num_steps,
-            mpi_handler=mpi_handler,
-            pop_control_method=pop_control_method,
-            min_weight=min_weight,
-            max_weight=max_weight,
-            reconfiguration_frequency=reconfiguration_frequency,
+            nwalkers,
             verbose=verbose
         )
 
@@ -104,10 +92,10 @@ class UHFWalkers(BaseWalkers):
         )
 
         self.Ghalfa = numpy.zeros(
-            shape=(self.nwalkers, nup, self.nbasis), dtype=numpy.complex128
+            shape=(self.nwalkers, self.nup, self.nbasis), dtype=numpy.complex128
         )
         self.Ghalfb = numpy.zeros(
-            shape=(self.nwalkers, ndown, self.nbasis),
+            shape=(self.nwalkers, self.ndown, self.nbasis),
             dtype=numpy.complex128,
         )
 
@@ -208,31 +196,17 @@ class UHFWalkersParticleHole(UHFWalkers):
         self,
         initial_walker,
         nup, ndown, nbasis,
-        num_walkers_local,
-        num_walkers_global,
-        num_steps,
-        ndets,
+        nwalkers,
         mpi_handler=None,
-        pop_control_method="pair_branch",
-        min_weight=0.1,
-        max_weight=4,
-        reconfiguration_frequency=50,
         verbose=False
     ):
         super().__init__(
             initial_walker,
             nup, ndown, nbasis,
-            num_walkers_local,
-            num_walkers_global,
-            num_steps,
+            nwalkers,
             mpi_handler=mpi_handler,
-            pop_control_method=pop_control_method,
-            min_weight=min_weight,
-            max_weight=max_weight,
-            reconfiguration_frequency=reconfiguration_frequency,
             verbose=verbose
         )
-        self.ndets = ndets
         if verbose:
             print(
                 "# Initial overlap of walker with trial wavefunction: {:13.8e}".format(
@@ -256,6 +230,7 @@ class UHFWalkersParticleHole(UHFWalkers):
             dtype=numpy.complex128,
         )  # reference 1-GF
     def build(self,trial):
+        self.num_dets = trial.num_dets
         self.CIa = numpy.zeros(
             shape=(self.nwalkers, trial.nact, trial.nocc_alpha),
             dtype=numpy.complex128,
@@ -280,55 +255,37 @@ class UHFWalkersParticleHoleNaive(UHFWalkersParticleHole):
         self,
         initial_walker,
         nup, ndown, nbasis,
-        num_walkers_local,
-        num_walkers_global,
-        num_steps,
-        ndets,
+        nwalkers,
         mpi_handler=None,
-        pop_control_method="pair_branch",
-        min_weight=0.1,
-        max_weight=4,
-        reconfiguration_frequency=50,
         verbose=False
     ):
         super().__init__(
             initial_walker,
             nup, ndown, nbasis,
-            num_walkers_local,
-            num_walkers_global,
-            num_steps,
-            ndets,
+            nwalkers,
             mpi_handler=mpi_handler,
-            pop_control_method=pop_control_method,
-            min_weight=min_weight,
-            max_weight=max_weight,
-            reconfiguration_frequency=reconfiguration_frequency,
             verbose=verbose
         )
-        self.ndets = ndets
-        if verbose:
-            print(
-                "# Initial overlap of walker with trial wavefunction: {:13.8e}".format(
-                    self.ot.real
-                )
-            )
+
+    def build(self,trial):
+        self.num_dets = trial.num_dets
         # TODO: RENAME to something less like weight
         # This stores an array of overlap matrices with the various elements of
         # the trial wavefunction.
         self.det_weights = numpy.zeros(
-            (self.nwalkers, self.ndets), dtype=numpy.complex128
+            (self.nwalkers, self.num_dets), dtype=numpy.complex128
         )
         self.det_ovlpas = numpy.zeros(
-            (self.nwalkers, self.ndets), dtype=numpy.complex128
+            (self.nwalkers, self.num_dets), dtype=numpy.complex128
         )
         self.det_ovlpbs = numpy.zeros(
-            (self.nwalkers, self.ndets), dtype=numpy.complex128
+            (self.nwalkers, self.num_dets), dtype=numpy.complex128
         )
 
         self.Gia = numpy.zeros(
             shape=(
                 self.nwalkers,
-                self.ndets,
+                self.num_dets,
                 self.nbasis,
                 self.nbasis,
             ),
@@ -337,21 +294,20 @@ class UHFWalkersParticleHoleNaive(UHFWalkersParticleHole):
         self.Gib = numpy.zeros(
             shape=(
                 self.nwalkers,
-                self.ndets,
+                self.num_dets,
                 self.nbasis,
                 self.nbasis,
             ),
             dtype=numpy.complex128,
         )
         self.Gihalfa = numpy.zeros(
-            shape=(self.nwalkers, self.ndets, self.nup, self.nbasis),
+            shape=(self.nwalkers, self.num_dets, self.nup, self.nbasis),
             dtype=numpy.complex128,
         )
         self.Gihalfb = numpy.zeros(
-            shape=(self.nwalkers, self.ndets, self.ndown, self.nbasis),
+            shape=(self.nwalkers, self.num_dets, self.ndown, self.nbasis),
             dtype=numpy.complex128,
         )
-    def build(self,trial):
         self.CIa = numpy.zeros(
             shape=(self.nwalkers, self.nbasis, self.nbasis),
             dtype=numpy.complex128,
